Delete errorneous empty content length header (#53843)
fixes https://github.com/vercel/next.js/issues/53822

PR https://github.com/vercel/next.js/pull/53368 introduced a regression causing fetch with custom headers to fail in safari.
There are known issues with uncidi when a content-length: 0 header exists.

When performing a custom fetch in safari, this header is present.
When performing a custom fetch in chrome, this header is not present.




Co-authored-by: Zack Tanner <1939140+ztanner@users.noreply.github.com>