chore: add extra error info for rsc info helper (#54609)
When I was working on #54199 , if the file path is wrong the next app build error only shows "can't read file from ...", it took me hours to figure out where it actually fails in next build command. Adding an prefix for the error message so that could identify it easier.

Also clean up other unused comments and incorrect named variables